pyproject.toml
[tool.poetry]
name = "CodeReviewAI"
version = "0.1.0"
description = ""
authors = ["YevheniiMelnikov <evgeniy@workmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^74.1.2"
fastapi = "^0.115.3"
httpx = "^0.27.2"
redis = "^5.1.1"
pytest = "^8.3.3"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
openai = "^1.52.2"
loguru = "^0.7.2"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.3"
mypy = "^1.11.2"
mypy-extensions = "^1.0.0"